gas opt

+++ packages/deployments/contracts/contracts/core/connext/facets/SwapAdminFacet.sol
@@ -123,7 +123,7 @@ contract SwapAdminFacet is BaseConnextFacet {
     if (_pooledTokens.length <= 1 || _pooledTokens.length > 32)
       revert SwapAdminFacet__initializeSwap_invalidPooledTokens();
 
-    uint8 numPooledTokens = uint8(_pooledTokens.length);
+    uint256 numPooledTokens = _pooledTokens.length;
 
     if (numPooledTokens != decimals.length) revert SwapAdminFacet__initializeSwap_decimalsMismatch();
